requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
openai==1.12.0
anthropic==0.18.1
aiohttp==3.9.3
aiofiles==23.2.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
